requirements.txt
python-ldap
python-dotenv
requests
mysql-connector-python